Tidy fieldtype
