+++ src/app/[locale]/(logged-out)/data-custodian/[dataCustodianId]/config.tsx
@@ -1,12 +1,5 @@
-export enum FieldType {
-    TEXT = "text",
-    LIST_TEXT = "list",
-    LIST_LINK = "list-link",
-    DATE = "date",
-    LINK = "link",
-    TAG = "tag",
-    WYSIWYG = "wysiwyg",
-}
+import { FieldType } from "@/interfaces/FieldType";
+
 export interface DataCustodianField {
     path: string;
     type: FieldType;
