+++ src/interfaces/FieldType.ts
@@ -1,0 +1,11 @@
+export enum FieldType {
+    TEXT = "text",
+    LIST_TEXT = "list",
+    LIST_LINK = "list-link",
+    LIST_TAG = "list-tag",
+    LIST_DATASETTYPE = "dataset-type-list",
+    DATE = "date",
+    TAG = "tag",
+    LINK = "link",
+    WYSIWYG = "wysiwyg",
+}
